svr/requirements.txt
flask
flask-cors
openai
python-dotenv
mcp
jira
httpx
starlette
uvicorn

